Sleep of Death patch
According to multiple sources, bad things happen when sleep is entered
at frequencies other than 800 MHz.

This patch locks the CPU at 800 prior to suspend, and returns the min/max
once suspend is exited.

This is based on the following patches, both of which achieve the same goal
for two different architectures:
https://github.com/Ninpo/ninphetamine/commit/da121dcb461f6eca31db3b20b274b39ef6b853d7
The above link was, however, setting the CPU to 500 instead of 800 in conflict with other
patch implementations, discussions with netarchy, and netarchy's later adjustment patch to
handle overclock kernels

https://github.com/Entropy512/linux_kernel_sgh-i997r/commit/67cebea21c4487b18527fa9133765d67d5972f6c
For the Samsung Infuse