mach-s5pv310: enable android ram console on I777
This allows us to get /proc/last_kmsg after a crash

This is a backport of the implementation found on many I9100 kernels.

In theory the one in the AT&T source drop is "cleaner" as it's a kernel runtime option,
but in practice, we can't change our runtime options so we need to hardcode it
on I777.  :(

Signed-off-by: Andrew Dodd <atd7@cornell.edu>