Revert "net: wireless: bcmdhd: Fix memory leak in case of dhd_bus_init() failure"
This reverts commit a85c0a3e46f192d94a9127017c31b83564832f6a.

If you take time to read the commit - it does nothing.  It adds a check to see if
reset_flag is true - but in all calls to dhdsdio_release_dongle(), the new
reset_flag is ALWAYS set to true - so this patch actually does nothing.

There may be another patch that depends on this one, but it's hidden in the bowels
of the internet.

Thanks to gokhanmoral of SiyahKernel for pointing this one out.

Signed-off-by:  Andrew Dodd <atd7@cornell.edu>