pm-hotplug:  Add enhanced hotplugging tunables
This allows the on/off thresholds to be tuned dependent on screen state
along with the hotplugging poll rate.

It also lets hotplugging get locked out at lower frequencies (default
800 MHz)

Credit for this goes to gokhanmoral of SiyahKernel - this is a rebase
of all of his pm-hotplug commits minus the Tegrak stuff which is
unnecessary with this code.  (The only thing Tegrak Second Core can
achieve which this code can't is forcing CPU1 on - we never want
to do that since that inhibits all deeper idle states.)

Slightly different logic is used for the CPU1 turnon frequency -
allow it to turn on when it equals the limit.

Signed-off-by: Andrew Dodd <atd7@cornell.edu>